+++ AlchemistCode/Cards/Uncommon/FreshBatch.cs
@@ -3,14 +3,14 @@ using BaseLib.Utils;
 using MegaCrit.Sts2.Core.Entities.Cards;
 using MegaCrit.Sts2.Core.GameActions.Multiplayer;
 
-namespace Alchemist.AlchemistCode.Cards.Common;
+namespace Alchemist.AlchemistCode.Cards.Uncommon;
 
 [CardTheme(CardTheme.Mix)]
 public class FreshBatch : AlchemistCard
 {
     protected internal override bool PlaysCastAnimation => false;
 
-    public FreshBatch() : base(1, CardType.Skill, CardRarity.Common, TargetType.Self)
+    public FreshBatch() : base(1, CardType.Skill, CardRarity.Uncommon, TargetType.Self)
     {
         WithCostUpgradeBy(-1);
         WithCards(1, 0);
@@ -19,7 +19,7 @@ public class FreshBatch : AlchemistCard
 
     protected override async Task OnPlay(PlayerChoiceContext choiceContext, CardPlay play)
     {
-        await Mixing.CreateChosen(choiceContext, Owner);
         await CommonActions.Draw(this, choiceContext);
+        await Mixing.CreateChosen(choiceContext, Owner);
     }
 }